move listProcesses into processes module
