+++ frontend/src/components/experimentRuns/ExperimentRunsDetailContent.js
@@ -7,7 +7,8 @@ const { TabPane } = Tabs;
 import AppPageHeader from "../AppPageHeader";
 import PageContent from "../PageContent";
 import TrackingFieldsContent from "../TrackingFieldsContent";
-import {get, listProcesses, listPropertyValues} from "../../modules/experimentRuns/actions";
+import {get, listPropertyValues} from "../../modules/experimentRuns/actions";
+import { list as listProcesses } from  "../../modules/processes/actions";
 import {withContainer} from "../../utils/withItem";
 import ProcessProperties from "../shared/ProcessProperties";
 import ExperimentRunsSamples from "./ExperimentRunsSamples";
