+++ frontend/src/components/processMeasurements/ProcessMeasurementsDetailContent.js
@@ -10,7 +10,8 @@ import AppPageHeader from "../AppPageHeader";
 import PageContent from "../PageContent";
 import ProcessProperties from "../shared/ProcessProperties";
 import TrackingFieldsContent from "../TrackingFieldsContent";
-import {listPropertyValues, listProcesses} from "../../modules/experimentRuns/actions";
+import {listPropertyValues} from "../../modules/experimentRuns/actions";
+import { list as listProcesses } from "../../modules/processes/actions";
 import {withSample} from "../../utils/withItem";
 import {get} from "../../modules/processMeasurements/actions";
 import { isProcessPropertiesLoaded } from "../../utils/actionsWait";
